# docs/swagger.yaml
basePath: /
definitions:
  domain.CreateTransactionMessage:
    properties:
      amount:
        type: number
      timestamp:
        description: 'Optional: can be filled in if not present'
        type: string
      transaction_type:
        description: '"bet" or "win"'
        type: string
      user_id:
        type: string
    type: object
  domain.Transaction:
    properties:
      amount:
        type: number
      timestamp:
        type: string
      transaction_type:
        type: string
      user_id:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API for managing user transactions via Kafka and PostgreSQL
  title: Transaction Manager API
  version: "1.0"
paths:
  /transactions:
    get:
      description: Get transactions by user ID and optional type
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Transaction type (bet|win)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Transaction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Enqueue a new transaction via Kafka
      parameters:
      - description: Transaction payload
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTransactionMessage'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new transaction
      tags:
      - transactions
schemes:
- http
swagger: "2.0"